Refactor analysis routines into nested loops
- Also swaps object_array with object_dict
- Adds SUFFICIENT_SCORE_THRESHOLD (unnamed prior) with default value 0.5
- Does away with extraneous string formatting previously introduced
- Slight changes to comments to match changes
- Score threshold check changed from '>' to '>='

The previous state of the routines for this operation were hacked
together to facilitate testing and review, but they were not ultimately
up to snuff and needed to be revamped.

The change from object_array to object_dict is to facilitate extensible
analysis approaches without necessitating modification of this code and
ultimately enable easy storage and retrieval of scores, filenames, etc.